Pr/845 (#1)
* btcjson: add btcd extension commands for mining addr control

This commit adds three new btcd extension commands: addminingaddr,
delminingaddr, and listminingaddrs. These new commands allows users to
dynamically add, remove and list the current set of mining addresses available
to btcd.

* rpcserver: implement addminingaddr, delminingaddr, listminingaddrs

This commit implements a new btcd extension RPC calls: addminingaddr,
delminingaddr, and listminingaddrs.  These RPC calls allows a caller to add,
remove and list mining addresses dynamically without restarting a running btcd
node.  Within btcd’s set of active mining addresses, all addresses must be
unique.  Incoming RPC calls that attempt to add duplicate generation addresses
are ignored. Similarly, any attempts to delete a non-existent mining address
will result in an error.

In order to implement this new RPC call in a thread-safe manner, a new
mutex has been added to synchronize access the `miningAddrs` attribute
within the `config` struct.

* mining/cpuminer: modify config to add new GetMiningAddr func

This commit modifies the existing `cpuminer.Config` struct to replace
with the `MiningAddrs` attribute and manual generation address
selection with a new function closure. When called, this function
closure returns a mining address to be used within the next generated
block. This modification adds a layer of abstraction over the storage
and selection logic for obtaining a new mining address.

* integration: add tests for addminingaddr, delminingaddr and listminingaddrs

* docs/json_rpc_api: document the new mining addr control RPCs

* config: fix bug in mining addr config parsing

This commit fixes a slight bug in the parsing of the supplied mining addresses
within the configuration. Before this commit, the code was checking the length
of `cfg.MiningAddrs` (the populated config value) instead of `cfg.miningAddrs`
the value in which the valid mining addresses are stored within.

* patch based on @davecgh's last review comments on #845